backend refactor to clean up projects
