+++ backend/config/urls.py
@@ -19,8 +19,6 @@ from django.urls import path
 
 from app import views, api_views
 from app.summer_2021_prototypes import views as summer_prototype_views
-from app.synthesize_polygons import views as synthesize_polygons_views
-from app.color_encoding_to_sound import views as color_encoding_to_sound_views
 
 urlpatterns = [
     # Django admin page
@@ -32,15 +30,15 @@ urlpatterns = [
     path('api/generate_instrument_2d/', api_views.generate_instrument_2d),
     path('api/gesture_to_sound/', api_views.gesture_to_sound),
     path('api/get_shape_analysis/', api_views.get_shape_analysis),
-    path('api/synthesize_polygon/', synthesize_polygons_views.synthesize_polygon_endpoint),
+    path('api/synthesize_polygon/', api_views.synthesize_polygons),
     path('api/playback_demo/', api_views.playback_demo),
 
     # View paths
     path('', views.index, name='index'),
     path('time-series/', views.time_series),
-    path('colors/', color_encoding_to_sound_views.base_page),
+    path('colors/', views.colors),
     path('gestures/', views.gestures),
-    path('polygons/', synthesize_polygons_views.synthesize_polygons),
+    path('polygons/', views.polygons),
     path('text-shape/', views.text_shape_analysis),
     path('numbers/', views.numbers),
     path('playback-demo/', views.playback_demo),
@@ -57,29 +55,29 @@ summer_prototype_urlpatterns = [
 urlpatterns += summer_prototype_urlpatterns
 
 
-################################################################################
-# Examples of how to implement these
-################################################################################
-example_urlpatterns = [
-    # Regular GET endpoint
-    # Creates a new view at the route /example/ and calls the view function at views.example
-    path('example/', views.example, name='example'),
-
-    # GET endpoint with a parameter.
-    # The part of the URL inside of <> gets parsed into a keyword argument and sent to the
-    # view function. See views.example_id: the keyword argument in the URL
-    # matches the keyword argument to the view function.
-    path('example/<int:example_id_arg>/', views.example_id, name='example_id'),
-
-    # GET API endpoint with a parameter.
-    # This API call takes arguments just like the endpoint above
-    path('api/example/<int:api_example_id_arg>/', api_views.get_example),
-
-    # POST API endpoint with a parameter.
-    # This API call doesn't register its arguments here, but take a look at api_views.post_example
-    # This endpoint has to be called via a POST request, with the data payload that will end up
-    # in request.data contained in the body of the request.
-    path('api/post_example/', api_views.post_example),
-]
-
-urlpatterns += example_urlpatterns
+# ################################################################################
+# # Examples of how to implement these
+# ################################################################################
+# example_urlpatterns = [
+#     # Regular GET endpoint
+#     # Creates a new view at the route /example/ and calls the view function at views.example
+#     path('example/', views.example, name='example'),
+#
+#     # GET endpoint with a parameter.
+#     # The part of the URL inside of <> gets parsed into a keyword argument and sent to the
+#     # view function. See views.example_id: the keyword argument in the URL
+#     # matches the keyword argument to the view function.
+#     path('example/<int:example_id_arg>/', views.example_id, name='example_id'),
+#
+#     # GET API endpoint with a parameter.
+#     # This API call takes arguments just like the endpoint above
+#     path('api/example/<int:api_example_id_arg>/', api_views.get_example),
+#
+#     # POST API endpoint with a parameter.
+#     # This API call doesn't register its arguments here, but take a look at api_views.post_example
+#     # This endpoint has to be called via a POST request, with the data payload that will end up
+#     # in request.data contained in the body of the request.
+#     path('api/post_example/', api_views.post_example),
+# ]
+#
+# urlpatterns += example_urlpatterns
